Update tests to reflect new controller structure
